Feature: adapt simple docker setup for development
This should provide a relatively simple docker setup with a few minimal tweaks
on the original version described in the wiki (see README). Basically, we
are installing plugins in interactive mode during the build of the container,
then remounting them at runtime in order to be able to save any migrations to
the local environment (outside the container).

Tested with netbox_access_lists.